analyzer/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
deprecation==2.1.0
distro==1.9.0
et_xmlfile==2.0.0
frozenlist==1.5.0
gotrue==2.11.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.0.1
idna==3.10
jiter==0.8.2
lxml==5.3.0
multidict==6.1.0
numpy==2.2.2
openai==1.59.8
openpyxl==3.1.5
packaging==24.2
pillow==11.1.0
postgrest==0.19.1
propcache==0.2.1
pydantic==2.10.5
pydantic_core==2.27.2
PyMuPDF==1.25.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
realtime==2.1.0
requests==2.32.3
resend==2.5.1
six==1.17.0
sniffio==1.3.1
storage3==0.11.0
StrEnum==0.4.15
supabase==2.11.0
supafunc==0.9.2
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
websockets==13.1
yarl==1.18.3
pytz==2023.3.post1
python-dotenv==1.0.1
